elfmalloc: Stop wasting virtual memory.
This change adds support for 2 PageAlloc-backed groups of slabs in
elfmalloc. This is reminiscent of the techniques used in the rust
version of the allocator, but it still holds "medium" objects in slabs.

In order to support this, a new AllocType variant was added and
PageAlloc's now support ensuring some page is aligned to a higher
alignment. This is done by using a second MemorySource as we must ensure
that the "round down" operation is safe.

This currently makes the code a bit harder to read; cleanups will
hopefully follow.