bsalloc: utilize all the allocated memory from the caches
The underlying MapAlloc allocator will round up the allocation
(to the nearest page), but the cache was only servicing requests
smaller than the allocation that was requested, rather than the
allocation actually given.

Use `usable_size()` to use the allocated size, rather than the
requested size. In practice (at least on Linux), this changes
small objects from 1k to 4k, and leaves the size of large objects
unchanged.