bsalloc: improve rng()
rng() is a multiple of the alignment of `usize` (in particular of the
memory address of an `usize`). When taking this modulo 4 (and 8, on
64 bit), the result is always 0. This means that at most two entries
in the caches are utilized.

Divide the output by the alignment of `usize`, which preserves the
overall entropy, but moves it into the lower bits (which are the ones
that are used).