bsalloc: improve rng() further
Further to the previous patch, the address taken as `seed`, is a
fixed offset from the stack frame alignment. On OSX this is fixed
to 16 bytes, and on other systems 16 bytes is the default.

Divide `seed` by 16, rather than simply the alignment of `usize`.
As before, this does not reduce the overall entropy (assuming default
stack-frame alignment), but moves more entropy into the lower bits.